Removed unnecessary include from TextBuffer.idl Changed highlight_screen's parameters from float to double

+++ Berlin/include/Berlin/Console.hh
@@ -113,9 +113,9 @@ public:
   //. you do not pass a color it defaults to red.
   virtual void highlight_screen(Warsaw::Coord, Warsaw::Coord,
 				Warsaw::Coord, Warsaw::Coord,
-				float red = 1.0,
-				float green = 0.0,
-				float blue = 0.0) = 0;
+				double red = 1.0,
+				double green = 0.0,
+				double blue = 0.0) = 0;
 
   template <typename T>
   T *get_extension(const std::string &id)
