+++ SDL/src/Console.hh
@@ -126,9 +126,9 @@ public:
   virtual void activate_autoplay();
   virtual void highlight_screen(Warsaw::Coord, Warsaw::Coord,
 				Warsaw::Coord, Warsaw::Coord,
-				float red = 1.0,
-				float green = 0.0,
-				float blue = 0.0);
+				double red = 1.0,
+				double green = 0.0,
+				double blue = 0.0);
   virtual void wakeup();
 
   void add_drawable(SDL::Drawable *);
